pyproject.toml
[project]
name = "pde-sim"
version = "0.1.0"
description = "2D PDE Simulation Dataset Generator for Vision-Language Models"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "h5py>=3.15.1",
    "matplotlib>=3.10.8",
    "numba>=0.63.1",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "py-pde>=0.50.0",
    "pyfftw>=0.15.1",
    "pytest>=9.0.2",
    "scipy>=1.16.3",
    "sympy>=1.14.0",
    "tqdm>=4.67.1",
    "pyyaml>=6.0",
]

[project.scripts]
pde-sim = "pde_sim.__main__:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
